Move the DevTools toggling to a FAB
Instead of requiring the user to navigate to the settings page to toggle
DevTools, make it togglable by an always-visible floating action button. It's
only visible in development builds, mind you.

Fixes #47.
